Util::NoticeMultiLines, Util::NoticeOnEachChannel を IrcPlugin に統合
Util::NoticeOnEachChannel#notice_on_each_channel メソッドに相当する機能を、IrcPlugin#send_notice メソッドに統合した。
これで、NOTICE コマンド関連の全てのメソッドは、IrcPlugin#send_notice メソッドだけを呼び出せば良いようになった。
